# backend/empresa/target/generated-sources/openapi/src/main/java/com/dasad/empresa/api/ChatApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dasad.empresa.api;

import com.dasad.empresa.model.ChatRequestDTO;
import com.dasad.empresa.model.ChatResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-15T21:23:23.313933800Z[Europe/Lisbon]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "chat", description = "Operações relacionadas a chat")
public interface ChatApi {

    /**
     * POST /chat/generate : Generate endpoint
     * Endpoint para gerar resposta baseada no modelo e prompt fornecidos.
     *
     * @param chatRequestDTO  (required)
     * @return Sucesso (status code 200)
     */
    @Operation(
        operationId = "generate",
        summary = "Generate endpoint",
        description = "Endpoint para gerar resposta baseada no modelo e prompt fornecidos.",
        tags = { "chat" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ChatResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/chat/generate",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ChatResponseDTO> generate(
        @Parameter(name = "ChatRequestDTO", description = "", required = true) @Valid @RequestBody ChatRequestDTO chatRequestDTO
    );

}